Move NoteEncryption tests to gtest suite.
